fixed bugs
the algorithm should be working better now. I fixed a lot of problems
with the recursive step. atm it is really not efficient for sudoku-sized
constraint matrix. In fact, I let it run for a long time and it did not
finish, so either something is wrong or it's just inefficient.